정리

+++ src/main/java/sseonmi/step1/Lotto.java
@@ -5,13 +5,13 @@ import java.util.Set;
 public class Lotto {
     private Set<Integer> randomNumberList;
 
-    private long matchCount;
+    private int matchCount;
 
     public Set<Integer> getRandomNumberList() {
         return randomNumberList;
     }
 
-    public long getMatchCount() {
+    public int getMatchCount() {
         return matchCount;
     }
 
@@ -19,7 +19,7 @@ public class Lotto {
         this.randomNumberList = randomNumberList;
     }
 
-    public void setMatchCount(long matchCount) {
+    public void setMatchCount(int matchCount) {
         this.matchCount = matchCount;
     }
 }
